requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pyyaml>=6.0
matplotlib>=3.7.0
pandas>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
ollama>=0.3.0
